Adding derive clone to the csv cell data type and using usize.

+++ rust/src/lib.rs
@@ -2,6 +2,7 @@ use wasm_bindgen::prelude::*;
 use csv::ReaderBuilder;
 use csv::StringRecordsIter;
 
+#[derive(Clone)]
 enum CsvCellDataType {
   Int,
   Float,
@@ -53,7 +54,7 @@ pub fn create_sql_schema_from_csv(
 fn create_csv_record_data_types_and_max_text_lengths(
   number_of_columns: usize,
   records: StringRecordsIter
-) -> (Vec<crate::CsvCellDataType>, Vec<u32>) {
+) -> (Vec<crate::CsvCellDataType>, Vec<usize>) {
   let record_data_types: Vec<crate::CsvCellDataType> =
     vec![crate::CsvCellDataType::Empty; number_of_columns];
